% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boruta_method.R
\name{boruta_method}
\alias{boruta_method}
\title{boruta_method}
\usage{
boruta_method(data, ntree, mtry, selected, ...)
}
\arguments{
\item{data}{Training data of class \code{data.frame} or \code{matrix}}

\item{ntree}{\code{integer(1)} number of tree.}

\item{mtry}{\code{integer(1)} number of Variables randomly chosen at each split}

\item{selected}{variables selected}

\item{\ldots}{further arguments passed to \code{\link[=ranger]{ranger()}}.}
}
\value{
\code{ks} object, see \code{\link[=knowledge_slanted]{knowledge_slanted()}} for details.
}
\description{
Boruta is an all relevant feature selection wrapper algorithm,
\emph{boruta_method} is a wrapper for knowledge-slanted random forest implementation.
It tries to pick all the important dataset features
}
\details{
Runs a feature selection function
}
\references{
Biological knowledge-slanted random forest approach
for the classification of calcified aortic valve stenosis
10.1186/s13040-021-00269-4
}
\seealso{
\code{\link[=knowledge_slanted]{knowledge_slanted()}}
}
