% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_slanted.R
\name{knowledge_slanted}
\alias{knowledge_slanted}
\alias{slanted}
\alias{ks}
\title{knowledge_slanted}
\usage{
knowledge_slanted(train, ntree, mtry, weights, ...)
}
\arguments{
\item{train}{Training data of class \code{data.frame}, \code{matrix}}

\item{ntree}{Number of trees.}

\item{mtry}{Number of variables to possibly split at in each node.}

\item{weights}{Numeric vector with weights between 0 and 1}

\item{\ldots}{further arguments passed to \code{\link[=ranger]{ranger()}}.}
}
\description{
knowledge_slanted
}
\seealso{
\code{\link[=ranger]{ranger()}}
}
