% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksborutahits.R
\name{ksborutahits}
\alias{ksborutahits}
\title{ksborutahits}
\usage{
ksborutahits(
  iter = 100,
  method = "bonferroni",
  p_value = 0.05,
  data = NULL,
  ntree,
  mtry,
  weights,
  runs = 20,
  trace = 0,
  ...
)
}
\arguments{
\item{iter}{maximum number of iterations}

\item{method}{adjust method to tune \code{\link[stats:p.adjust]{stats::p.adjust()}}}

\item{p_value}{\code{number} probability of obtaining test results}

\item{data}{Training data of class \code{data.frame} or \code{matrix}}

\item{ntree}{\code{integer(1)} number of tree.}

\item{mtry}{\code{integer(1)} number of Variables randomly chosen at each split}

\item{weights}{numeric vector with weights between 0 and 1.}

\item{runs}{minimum number of iterations. Default value should be used.}

\item{trace}{verbose. 0 means no tracing, 1 means decision about each attribute, 2 means the same as 1, plus reporting step by step}

\item{\ldots}{further arguments passed to \code{ksboruta}}
}
\description{
ksborutahits to  replicate \emph{ksboruta} method
}
\examples{
# basic usage of ksborutahits
data <- simulated2$train
mtry <- sqrt(length(data))
weights <- simulated2$weights
ksborutahits(iter=50, data=data, ntree = 100, mtry = mtry, weights = weights, trace=TRUE)

}
