% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kslboruta-package.R
\docType{package}
\name{kslboruta-package}
\alias{kslboruta-package}
\alias{_PACKAGE}
\alias{knowledge-slanted-boruta-package}
\alias{ks-boruta-package}
\alias{knowledgeslantedboruta-package}
\title{kslboruta: Modified knowledge-slanted implementation to support boruta method}
\description{
The knowledge-slanted random forest implementation is modified to support boruta method. Additional helper functions are also provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ErikaCantor/knowledge-slanted-boruta}
  \item Report bugs at \url{https://github.com/ErikaCantor/knowledge-slanted-boruta/issues/}
}

}
\author{
\strong{Maintainer}: Erika Cantor \email{erika.cantor@postgrado.uv.cl} (\href{https://orcid.org/0000-0003-3320-6032}{ORCID})

}
\keyword{internal}
