% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksmetrics.R
\name{ksmetrics}
\alias{ksmetrics}
\alias{metrics}
\title{ksmetrics}
\usage{
ksmetrics(ksobject, data = NULL, predictions, ...)
}
\arguments{
\item{ksobject}{A knowledge_slanted object.}

\item{data}{test data}

\item{predictions}{aggregated predictions for all trees}

\item{\ldots}{further arguments passed to \code{\link[pROC:roc]{pROC::roc()}}.}
}
\description{
Compute Area Under the Curve (AUC).
ROC is a probability curve and AUC represents the degree or measure of separability
}
\seealso{
\code{\link[=knowledge_slanted]{knowledge_slanted()}}

\code{\link[=kspredict]{kspredict()}}
}
