% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kspredict.R
\name{kspredict}
\alias{kspredict}
\alias{predict}
\alias{ksp}
\title{kspredict}
\usage{
kspredict(ksobject, data = NULL, performance = FALSE, ...)
}
\arguments{
\item{ksobject}{A knowledge_slanted object.}

\item{data}{test data}

\item{\ldots}{further arguments passed to \code{\link[=predict]{predict()}}.}
}
\description{
Prediction with test data and a knowledge_slanted result.
}
\seealso{
\code{\link[=knowledge_slanted]{knowledge_slanted()}}

\code{\link[=ksmetrics]{ksmetrics()}}
}
