% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kswalk.R
\name{kswalk}
\alias{kswalk}
\title{random walk prior}
\usage{
kswalk(data, graph, seednodes, ...)
}
\arguments{
\item{data}{\code{data.frame}}

\item{graph}{\code{data.frame}}

\item{seednodes}{vector with the numbers of the columns of features that are seed nodes}

\item{\ldots}{further arguments passed to \code{random.walk}}
}
\value{
\verb{random walk} object
}
\description{
'kswalk' performs a random walk with restart in order to obtain the selection probability of each feature for the knowledged-slanted random forest.
}
