% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pulldata}
\alias{pulldata}
\title{Pull trait simulated and test data
This function read trait data}
\usage{
pulldata(name = NULL)
}
\arguments{
\item{name}{the dataset to source.}
}
\value{
a raw dataset available in R
}
\description{
Pull trait simulated and test data
This function read trait data
}
\examples{

# to get a list of all available data within the package
pulldata()

# to import a dataset called _simulated1_

}
\seealso{
Other simulated_data: 
\code{\link{simulated1_raw}},
\code{\link{simulated2_raw}}
}
\author{
Erika Cantor
}
\concept{simulated_data}
