% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_prior.R
\name{rw_prior}
\alias{rw_prior}
\title{random walk prior}
\usage{
rw_prior(data, graph, seednodes, ...)
}
\arguments{
\item{data}{\code{data.frame}}

\item{seednodes}{\code{data.frame}}
}
\value{
\verb{random walk} object
}
\description{
'rw_prior' perform random walk with restart
}
\seealso{
\code{\link{random.walk}}
}
