% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated1}
\alias{simulated1}
\title{ksldata knowledge_slanted data}
\format{
An object of class \code{list} of length 5.
}
\source{
\url{https://github.com/ErikaCantor/knowledge-slanted-boruta/data/ksldata1.rda}
}
\usage{
simulated1
}
\description{
A dataset containing train and test simulated values
}
\author{
Erika Cantor

A large list with two data frames
\describe{
\item{train}{dataframe with 30 obs. of 5001 variables}
\item{seed}{numeric array with seed values}
\item{weights}{numeric array with seed values}
\item{beta}{numeric array with seed values}
\item{id.p}{numeric array with seed values}
}
}
\keyword{datasets}
