% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{simulated1_raw}
\alias{simulated1_raw}
\title{Simulated train-test data set}
\format{
A large list
\describe{
\item{train}{dataframe with 27 obs. of 1001 variables}
\item{test}{dataframe with 12 obs. of 1001 variables}
}
}
\source{
\url{https://github.com/ErikaCantor/knowledge-slanted-boruta/data/simulated1.RData}
}
\usage{
pulldata(simulated1)
}
\description{
A dataset containing train and test simulated values
}
\examples{
pulldata(simulated1)
}
\seealso{
Other simulated_data: 
\code{\link{pulldata}()},
\code{\link{simulated2_raw}}
}
\author{
Erika Cantor
}
\concept{simulated_data}
