% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated2}
\alias{simulated2}
\title{ksdata train-test data set}
\format{
A large list
\describe{
\item{train}{A dataframe with 27 obs. of 1001 variables}
\item{test}{A dataframe with 12 obs. of 1001 variables}
}
}
\source{
\url{https://github.com/ErikaCantor/knowledge-slanted-boruta/data/ksldata2.rda}
}
\usage{
simulated2
}
\description{
Small simulated data set where RF Approach can be performed
}
\author{
Erika Cantor
}
\keyword{datasets}
