% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{simulated2_raw}
\alias{simulated2_raw}
\title{Simulated knowledge_slanted}
\format{
A large list
\describe{
\item{train}{dataframe with 30 obs. of 5001 variables}
\item{seed}{numeric array with seed values}
\item{weights}{numeric array with seed values}
\item{beta}{numeric array with seed values}
\item{id.p}{numeric array with seed values}
}
}
\source{
\url{https://github.com/ErikaCantor/knowledge-slanted-boruta/data/simulated2.RData}
}
\usage{
pulldata(simulated2)
}
\description{
Small simulated data sets where RF Approach can be performed
}
\examples{
pulldata(simulated2)
}
\seealso{
Other simulated_data: 
\code{\link{pulldata}()},
\code{\link{simulated1_raw}}
}
\author{
Erika Cantor
}
\concept{simulated_data}
