% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_slanted.R
\name{knowledge_slanted}
\alias{knowledge_slanted}
\title{knowledge_slanted}
\usage{
knowledge_slanted(data = NULL, ntree, mtry, weights, ...)
}
\arguments{
\item{data}{Training data of class \code{data.frame}, \code{matrix}. The result variable must be the first column of \code{data.frame} and labeled \code{y}.}

\item{ntree}{Number of trees.}

\item{mtry}{Number of variables to possibly split at in each node.}

\item{weights}{Numeric vector with weights between 0 and 1}

\item{\ldots}{further arguments to be passed to or from other methods especially
\itemize{
\item \code{test = data.frame}, to predict the values based on the previous data
\item \code{do.metrics = TRUE}, to list different metrics that were added to your model (error S E auc F1)
}}
}
\description{
 This function implements the Knowledge-slanted random forest (RF) for classification tasks.
}
\details{
The knowledge-slanted RF is a combination of two stages as an 
  attempt to implement a knowledge-guided supervised learning 
  approach based on the conventional RF. The weights can be given by the user or previously obtained by the kswalk function from a 
  protein protein interaction network or graph.
  
}
\references{
Leo Breiman. Random forests. Machine learning, 45(1):5–32, 2001.
  
Erika Cantor, Rodrigo Salas, Harvey Rosas, and Sandra Guauque-Olarte.
Biological knowledge-slanted random forest approach for the classification
of calcified aortic valve stenosis. BioData Mining, 14(1):1–11, 2021.
}
\seealso{
\code{\link[=ranger]{ranger()}}

\code{\link[=kspredict]{kspredict()}}

\code{\link[=ksmetrics]{ksmetrics()}}
}
