% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kswalk.R
\name{kswalk}
\alias{kswalk}
\title{random walk prior}
\usage{
kswalk(data, graph, seednodes, ...)
}
\arguments{
\item{data}{\code{data.frame}}

\item{graph}{\code{data.frame}}

\item{seednodes}{\code{data.frame}}

\item{\ldots}{further arguments passed to \code{random.walk}}
}
\value{
\verb{random walk} object
}
\description{
'kswalk' perform random walk with restart
}
